backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Cache
redis[hiredis]==5.1.1

# Data processing
polars==1.9.0
pandas==2.2.3
duckdb==1.1.1
pyarrow==17.0.0

# Statistics
scipy==1.14.1
numpy==2.1.2

# Synthetic data
faker==30.3.0

# Config
pyyaml==6.0.2
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.2

# Reports
jinja2==3.1.4
weasyprint==62.3
matplotlib==3.9.2

# Email / scheduling
apscheduler==3.10.4
aiosmtplib==3.0.2

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
structlog==24.4.0
